requirements.txt
certifi==2017.11.5
chardet==3.0.4
gevent==1.2.2
greenlet==0.4.12
idna==2.6
requests==2.18.4
six==1.11.0
urllib3==1.22
websocket-client==0.45.0
